Initial stab at tracing
As this is very much work in progress, here are some quick notes:
- demo/trace_model.py has the current state

Lots of todos:
- there are warnings to be investigated, in particularly for loops and
  boxlist sizes
- it needs some PyTorch JIT fixes
- clean up, including reverting unneeded changes
- round off the displaying
- do a C++ app
- try a different model/config

+++ maskrcnn_benchmark/csrc/ROIAlign.h
@@ -10,10 +10,10 @@
 // Interface for Python
 at::Tensor ROIAlign_forward(const at::Tensor& input,
                             const at::Tensor& rois,
-                            const float spatial_scale,
-                            const int pooled_height,
-                            const int pooled_width,
-                            const int sampling_ratio) {
+                            const double spatial_scale,
+                            const int64_t pooled_height,
+                            const int64_t pooled_width,
+                            const int64_t sampling_ratio) {
   if (input.type().is_cuda()) {
 #ifdef WITH_CUDA
     return ROIAlign_forward_cuda(input, rois, spatial_scale, pooled_height, pooled_width, sampling_ratio);
